refactor: cargo clippy auto-fix + FFI 函数指针抽 type alias
阶段 0 整理（作为后续 idiomatic Rust 重构的起点）：

- cargo clippy --fix 自动修的 idiomatic 改进：
  * Range::contains / div_ceil / copy_from_slice / Default impl /
    or_default / 删除冗余 `as` cast / needless borrow 等
- src/ffi/mod.rs: MyPaintSurface vtable 里 5 个长函数指针
  抽成具名 type alias (MyPaintSurfaceDrawDabFn 等)，可读性 + 后续
  内部代码引用更方便。Option<fn> ABI 不变，C 端兼容性零影响。

所有 46 tests 通过，--all-targets clippy correctness 通过。

+++ examples/basic_paint.rs
@@ -56,7 +56,7 @@ fn main() {
             pressure,
             0.0,
             0.0,
-            dt as f64,
+            dt,
             1.0,
             0.0,
             0.0,
